+++ src/com/rian/difficultycalculator/math/Vector2.java
@@ -9,14 +9,14 @@ public class Vector2 {
     /**
      * The X position of the vector.
      */
-    public double x;
+    public float x;
 
     /**
      * The Y position of the vector.
      */
-    public double y;
+    public float y;
 
-    public Vector2(double value) {
+    public Vector2(float value) {
         this(value, value);
     }
 
@@ -24,7 +24,7 @@ public class Vector2 {
         this(pointF.x, pointF.y);
     }
 
-    public Vector2(double x, double y) {
+    public Vector2(float x, float y) {
         this.x = x;
         this.y = y;
     }
@@ -78,8 +78,8 @@ public class Vector2 {
     /**
      * The length of this vector.
      */
-    public double getLength() {
-        return Math.hypot(x, y);
+    public float getLength() {
+        return (float) Math.hypot(x, y);
     }
 
     /**
@@ -88,7 +88,7 @@ public class Vector2 {
      * @param vec The other vector.
      * @return The dot product of both vectors.
      */
-    public double dot(Vector2 vec) {
+    public float dot(Vector2 vec) {
         return x * vec.x + y * vec.y;
     }
 
@@ -98,7 +98,7 @@ public class Vector2 {
      * @param scaleFactor The factor to scale the vector by.
      * @return The scaled vector.
      */
-    public Vector2 scale(double scaleFactor) {
+    public Vector2 scale(float scaleFactor) {
         return new Vector2(x * scaleFactor, y * scaleFactor);
     }
 
@@ -108,15 +108,15 @@ public class Vector2 {
      * @param vec The other vector.
      * @return The distance between this vector and the other vector.
      */
-    public double getDistance(Vector2 vec) {
-        return Math.hypot(x - vec.x, y - vec.y);
+    public float getDistance(Vector2 vec) {
+        return (float) Math.hypot(x - vec.x, y - vec.y);
     }
     
     /**
      * Normalizes the vector.
      */
     public void normalize() {
-        final double length = getLength();
+        final float length = getLength();
 
         x /= length;
         y /= length;
